requirements.txt
dnspython==2.2.1
pyspf==2.0.14
